# docs/api/OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: "KYC-AML Platform API"
  version: "2.0.0"
  description: |
    ## Plateforme KYC/AML — API de référence v2.0

    Cette API expose l'ensemble des fonctionnalités de la plateforme KYC-AML : gestion des clients,
    analyse transactionnelle temps réel, moteur de règles AML dynamiques, screening sanctions/PEP,
    gestion des alertes et des dossiers d'investigation, pKYC (Perpetual KYC) et production de
    rapports réglementaires (SAR, STR, goAML).

    ### Architecture
    L'API est construite sur **tRPC** exposé via Express. Les endpoints tRPC suivent la convention
    `POST /trpc/<router>.<procedure>` pour les mutations et `GET /trpc/<router>.<procedure>` pour
    les queries. Le webhook CBS est un endpoint REST standard.

    ### Authentification
    Tous les endpoints (sauf `/health`, `/metrics` et `/webhooks/transaction`) nécessitent un
    **Bearer JWT** valide obtenu via `auth.login`. Les tokens d'accès ont une durée de vie de
    **15 minutes**. Le refresh token est valide **7 jours**.

    ### Moteur AML
    Chaque transaction soumise déclenche synchroniquement le moteur de règles AML. Le score de
    risque (0–100) et les règles déclenchées sont retournés dans la réponse et stockés dans
    `risk_rules` (JSONB).

    ### Conformité réglementaire
    La plateforme supporte : TRACFIN (France), goAML v2 (GAFI/FATF), OFAC, EU Sanctions,
    UN Consolidated List, UK Sanctions List, OpenSanctions PEP.

    ### Contact & support
...
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenu via `POST /trpc/auth.login`.
        À inclure dans le header `Authorization: Bearer <accessToken>`.
        Durée de vie : 15 minutes. Renouveler via `auth.refresh`.
    WebhookHMAC:
      type: apiKey
      in: header
      name: x-webhook-signature
      description: |
        Signature HMAC-SHA256 du body brut de la requête.
        Format : `sha256=<hex-digest>`
        Calculée avec `WEBHOOK_SECRET` (variable d'environnement côté CBS).
        Exemple : `x-webhook-signature: sha256=a1b2c3d4e5f6...`

# ─── Tags ──────────────────────────────────────────────────────────────────────
...
paths:
...
  /trpc/alerts.resolve:
    post:
      tags: [alerts]
      operationId: alerts.resolve
      summary: "Résoudre ou fermer une alerte"
      description: |
        Clôture une alerte. Les valeurs de résolution possibles :
        - `FALSE_POSITIVE` : Alerte incorrecte, client légitime
        - `CLOSED` : Traitement terminé sans suite
        - `ESCALATED` : Escalade vers un dossier d'investigation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, resolution]
              properties:
                id:
                  type: integer
                resolution:
                  type: string
                  enum: [FALSE_POSITIVE, CLOSED, ESCALATED]
                notes:
                  type: string
                  description: "Commentaires de l'analyste"
      responses:
        '200':
          description: "Alerte résolue"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'

  # ══════════════════════════════════════════════════════════════════════════════
  # CASES
  # ══════════════════════════════════════════════════════════════════════════════
...
  /api/documents/upload:
    post:
      tags: [documents]
      operationId: documents.upload
      summary: "Uploader un document d'identité"
      description: |
        Upload multipart d'un document. Déclenche automatiquement :
        1. OCR (Tesseract) — extraction des champs et texte brut
        2. Analyse MRZ si applicable (passeport, carte d'identité)
        3. eKYC — vérification de cohérence et score de confiance
        Le traitement est synchrone. La réponse inclut les résultats OCR et eKYC.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, customerId, documentType]
              properties:
                file:
                  type: string
                  format: binary
                  description: "Fichier (PDF, JPEG, PNG — max 10 MB)"
                customerId:
                  type: integer
                documentType:
                  $ref: '#/components/schemas/DocumentType'
      responses:
        '200':
          description: "Document uploadé et traité"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  document:
                    $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '413':
          description: "Fichier trop volumineux (max 10 MB)"

  # ══════════════════════════════════════════════════════════════════════════════
  # WEBHOOKS
  # ══════════════════════════════════════════════════════════════════════════════
  /webhooks/transaction:
    post:
      tags: [transactions]
      operationId: webhooks.transaction
      summary: "Webhook CBS — réception de transaction"
      description: |
        Endpoint de réception des transactions émises par le CBS (Core Banking System).
        Ce endpoint ne requiert **pas** de Bearer JWT mais une **signature HMAC-SHA256**
        dans le header `x-webhook-signature`.

        ### Calcul de la signature
        ```
        HMAC-SHA256(WEBHOOK_SECRET, raw_body_bytes)
        ```
        Le header doit avoir la forme : `x-webhook-signature: sha256=<hex>`

        ### Idempotence
        Le champ `transactionId` est la clé d'idempotence. Une transaction avec un
        `transactionId` déjà existant retourne 200 sans retraitement.

        ### Traitement
        1. Vérification de la signature HMAC
        2. Déduplication (transactionId)
        3. Moteur AML synchrone (règles + ML si activé)
        4. Création alerte si score > seuil
        5. Retour du résultat complet
      security:
        - WebhookHMAC: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookTransactionPayload'
            examples:
              virement_entrant:
                summary: "Virement entrant"
                value:
                  transactionId: "CBS-2024-TX-089451"
                  customerId: 42
                  amount: "50000.00"
                  currency: "EUR"
                  transactionType: "TRANSFER"
                  channel: "API"
                  counterparty: "Shell Corp Ltd"
                  counterpartyCountry: "BVI"
                  counterpartyBank: "Offshore Trust Bank"
                  purpose: "Investment"
                  transactionDate: "2024-11-15T14:32:00Z"
      responses:
        '200':
          description: "Transaction traitée (ou déjà existante — idempotente)"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionWithAml'
        '400':
          description: "Payload invalide"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: "Signature HMAC invalide ou absente"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ══════════════════════════════════════════════════════════════════════════════
  # INFRASTRUCTURE
  # ══════════════════════════════════════════════════════════════════════════════
...